Improved functionality for high t-values and dof. E.g., for dof=5000, can correctly convert t-values up to ~40.

+++ +fpp/+util/convertTtoZ.m
@@ -8,8 +8,8 @@ function convertTtoZ(inputPath,outputPath,dof)
 [tVec,hdr] = fpp.util.readDataMatrix(inputPath);
 
 zVec = zeros(size(tVec));
-zVec(tVec>0) = -icdf('norm',tcdf(-tVec(tVec>0),dof),0,1);
-zVec(tVec<0) = icdf('norm',tcdf(tVec(tVec<0),dof),0,1); % Prevents p-values near 1 from rounding to 1
+zVec(tVec>0) = -icdf('norm',tcdf(-double(tVec(tVec>0)),dof),0,1);
+zVec(tVec<0) = icdf('norm',tcdf(double(tVec(tVec<0)),dof),0,1); % Prevents p-values near 1 from rounding to 1
 
 fpp.util.writeDataMatrix(zVec,hdr,outputPath);
